Fixed precision in Utils.getXPToLvl. Fixes #18.

+++ app/src/main/java/com/infonuascape/osrshelper/utils/Utils.java
@@ -9,16 +9,16 @@ import com.infonuascape.osrshelper.utils.players.PlayerSkills;
 
 public class Utils {
 
-	public static final float getXPToLvl(final int nextLvl){
+	public static final double getXPToLvl(final int nextLvl){
 		if(nextLvl == 100){
-			return 0f;
+			return 0;
 		}
 
-		float exp = 0f;
-		for(float i = 1 ; i <= nextLvl - 1 ; i++) {
-			exp += Math.floor(i + 300f * Math.pow(2f, i / 7f));
+		double exp = 0;
+		for(double i = 1 ; i <= nextLvl - 1 ; i++) {
+			exp += Math.floor(i + 300 * Math.pow(2, i / 7));
 		}
-		exp = (float) Math.floor(exp / 4f);
+		exp = (float) Math.floor(exp / 4);
 		
 		return exp;
 	}
